backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
groq==0.4.2
PyPDF2==3.0.1
python-docx==1.1.0
pylatex==1.4.2
python-dotenv==1.0.0
requests==2.31.0
supabase>=2.3.0